+++ sandbox/collections-0.2.0/src/main/java/vanilla/java/collections/impl/SelectedList.java
@@ -111,7 +111,7 @@ public class SelectedList<E> extends AbstractHugeCollection<E> {
     }
 
     @Override
-    public long partitionSize() {
+    public int partitionSize() {
       return SelectedList.this.partitionSize();
     }
